+++ tasks.py
@@ -127,7 +127,7 @@ def scan(c):
     """Integration testing: tests the scan command"""
     try:
         c.run(
-            "./cloudsplaining/bin/cli.py scan --input examples/files/example.json --exclusions-file examples/example-exclusions.yml --skip-open-report",
+            "./cloudsplaining/bin/cli.py scan --input-file examples/files/example.json --exclusions-file examples/example-exclusions.yml --skip-open-report",
             pty=True,
         )
     except UnexpectedExit as u_e: